pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "CANgSend"
version = "0.0.1"
authors = [
  { name = "Gnox79", email = "balusagnanesh16@gmail.com" }
]
description = "CANgSend is a Python package designed for sending and managing CAN (Controller Area Network) messages related to acceleration control. It provides a simple yet flexible API for creating and dispatching CAN messages, making it ideal for applications in automotive and industrial systems where CAN communication is essential."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["python-can", "argparse"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows"
]

[project.urls]
Homepage = "https://github.com/gnanesh-16/CANgSend"  
Issues = "https://github.com/gnanesh-16/CANgSend/issues"  